As a user, I can create a signed QR-code Fixes #15

+++ src/routes/project/[projectId]/+page.server.ts
@@ -1,4 +1,5 @@
 import { error, fail, redirect } from '@sveltejs/kit';
+import jwt from 'jsonwebtoken';
 import type { Actions, PageServerLoad } from './$types';
 
 export const load = (async ({ params: { projectId }, locals: { supabase, getSession } }) => {
@@ -61,7 +62,13 @@ export const actions = {
 			});
 		}
 	},
-	delete: async ({ params: { projectId }, locals: { supabase } }) => {
+	delete: async ({ params: { projectId }, locals: { supabase, getSession } }) => {
+		const session = await getSession();
+
+		if (!session) {
+			throw redirect(303, '/');
+		}
+
 		const { error } = await supabase.from('projects').delete().eq('id', projectId);
 
 		if (error) {
@@ -71,5 +78,34 @@ export const actions = {
 		}
 
 		throw redirect(303, `/`);
+	},
+	generateSignedLink: async ({ params: { projectId }, url, locals: { supabase, getSession } }) => {
+		const session = await getSession();
+
+		if (!session) {
+			throw redirect(303, '/');
+		}
+
+		const { data: profile } = await supabase
+			.from('profiles')
+			.select(`organization`)
+			.eq('id', session.user.id)
+			.single();
+
+		const payload: Record<string, unknown> = {
+			organization: profile?.organization,
+			project: projectId
+		};
+
+		// TODO: Change the secret to an env secret. Do NOT put this code into production like this.
+		const token = jwt.sign(payload, 'testymctestface');
+
+		const signedLink = `${url.origin}/public/project?signature=${token}`;
+
+		await new Promise((resolve) => setTimeout(resolve, 1000));
+
+		return {
+			signedLink
+		};
 	}
 } satisfies Actions;
